Fix build warnings with -Wall -Wextra -Werror
Address compiler warnings to enable strict warning flags:

- Move static halfsec variable from pmtr.h to pmtr.c where it's used
- Fix sign-compare warnings in utarray.h macros by using proper types
- Fix type-limits warnings in utarray_insert/inserta macros
- Add __attribute__((unused)) to sigs[] and rlimit_labels[] arrays
- Fix pointer signedness in job.c set_cpu function
- Change loop variables from int to size_t where comparing with sizeof()
- Remove unused variables and labels throughout
- Add CMake directive to suppress warnings for generated cfg.c
- Add Dockerfile for Linux build testing

All source files now compile cleanly with -Wall -Wextra -Werror.

+++ src/net.c
@@ -263,7 +263,7 @@ void close_sockets(pmtr_t *cfg) {
 
 /* report to all configured destinations */
 void report_status(pmtr_t *cfg) {
-  int rc;
+  ssize_t rc;
   time_t now = time(NULL);
 
   /* construct msg */
@@ -283,8 +283,8 @@ void report_status(pmtr_t *cfg) {
     rc = write(*fd,utstring_body(cfg->s),utstring_len(cfg->s));
     if (rc < 0 && errno != ECONNREFUSED) 
       syslog(LOG_INFO,"write error: %s", strerror(errno));
-    if (rc >= 0 && rc < utstring_len(cfg->s)) {
-      syslog(LOG_INFO,"incomplete write %d/%d", rc, utstring_len(cfg->s));
+    if (rc >= 0 && (size_t)rc < utstring_len(cfg->s)) {
+      syslog(LOG_INFO,"incomplete write %zd/%u", rc, utstring_len(cfg->s));
     }
   }
 }
